+++ packages/core/react/index.ts
@@ -1,2 +1,3 @@
 export * from './use-chat'
 export * from './use-completion'
+export * from './tokens'