+++ packages/core/react/use-completion.ts
@@ -60,6 +60,7 @@ export function useCompletion({
   id,
   initialCompletion = '',
   initialInput = '',
+  credentials,
   headers,
   body,
   onResponse,
@@ -80,16 +81,18 @@ export function useCompletion({
   const [abortController, setAbortController] =
     useState<AbortController | null>(null)
 
-  const extraMetadataRef = useRef<any>({
+  const extraMetadataRef = useRef({
+    credentials,
     headers,
     body
   })
   useEffect(() => {
     extraMetadataRef.current = {
+      credentials,
       headers,
       body
     }
-  }, [headers, body])
+  }, [credentials, headers, body])
 
   // Actual mutation hook to send messages to the API endpoint and update the
   // chat state.
@@ -120,6 +123,7 @@ export function useCompletion({
             ...extraMetadataRef.current.body,
             ...options?.body
           }),
+          credentials: extraMetadataRef.current.credentials,
           headers: {
             ...extraMetadataRef.current.headers,
             ...options?.headers