+++ packages/core/svelte/use-chat.ts
@@ -63,6 +63,7 @@ export function useChat({
   onResponse,
   onFinish,
   onError,
+  credentials,
   headers,
   body
 }: UseChatOptions = {}): UseChatHelpers {
@@ -118,7 +119,8 @@ export function useChat({
           ...headers,
           ...options?.headers
         },
-        signal: abortController.signal
+        signal: abortController.signal,
+        credentials
       }).catch(err => {
         // Restore the previous messages if the request fails.
         mutate(previousMessages)