+++ packages/core/tests/snapshots/openai-chat.ts
@@ -1,6 +1,4 @@
-// "Hello, world."
-
-export default [
+export const chatCompletionChunks = [
   {
     id: 'chatcmpl-7RyNSW2BXkOQQh7NlBc65j5kX8AjC',
     object: 'chat.completion.chunk',
@@ -60,5 +58,279 @@ export default [
     created: 1686901302,
     model: 'gpt-3.5-turbo-0301',
     choices: [{ delta: {}, index: 0, finish_reason: 'stop' }]
+  }
+]
+
+export const chatCompletionChunksWithFunctionCall = [
+  {
+    id: 'chatcmpl-7WBy19k4tnzMa0svAIAqkqeIaKZh8',
+    object: 'chat.completion.chunk',
+    created: 1687906853,
+    model: 'gpt-3.5-turbo-0613',
+    choices: [
+      {
+        index: 0,
+        delta: {
+          role: 'assistant',
+          content: null,
+          function_call: { name: 'get_current_weather', arguments: '' }
+        },
+        finish_reason: null
+      }
+    ]
+  },
+  {
+    id: 'chatcmpl-7WBy19k4tnzMa0svAIAqkqeIaKZh8',
+    object: 'chat.completion.chunk',
+    created: 1687906853,
+    model: 'gpt-3.5-turbo-0613',
+    choices: [
+      {
+        index: 0,
+        delta: { function_call: { arguments: '{\n' } },
+        finish_reason: null
+      }
+    ]
+  },
+  {
+    id: 'chatcmpl-7WBy19k4tnzMa0svAIAqkqeIaKZh8',
+    object: 'chat.completion.chunk',
+    created: 1687906853,
+    model: 'gpt-3.5-turbo-0613',
+    choices: [
+      {
+        index: 0,
+        delta: { function_call: { arguments: '"' } },
+        finish_reason: null
+      }
+    ]
+  },
+  {
+    id: 'chatcmpl-7WBy19k4tnzMa0svAIAqkqeIaKZh8',
+    object: 'chat.completion.chunk',
+    created: 1687906853,
+    model: 'gpt-3.5-turbo-0613',
+    choices: [
+      {
+        index: 0,
+        delta: { function_call: { arguments: 'location' } },
+        finish_reason: null
+      }
+    ]
+  },
+  {
+    id: 'chatcmpl-7WBy19k4tnzMa0svAIAqkqeIaKZh8',
+    object: 'chat.completion.chunk',
+    created: 1687906853,
+    model: 'gpt-3.5-turbo-0613',
+    choices: [
+      {
+        index: 0,
+        delta: { function_call: { arguments: '":' } },
+        finish_reason: null
+      }
+    ]
+  },
+  {
+    id: 'chatcmpl-7WBy19k4tnzMa0svAIAqkqeIaKZh8',
+    object: 'chat.completion.chunk',
+    created: 1687906853,
+    model: 'gpt-3.5-turbo-0613',
+    choices: [
+      {
+        index: 0,
+        delta: { function_call: { arguments: ' "' } },
+        finish_reason: null
+      }
+    ]
+  },
+  {
+    id: 'chatcmpl-7WBy19k4tnzMa0svAIAqkqeIaKZh8',
+    object: 'chat.completion.chunk',
+    created: 1687906853,
+    model: 'gpt-3.5-turbo-0613',
+    choices: [
+      {
+        index: 0,
+        delta: { function_call: { arguments: 'Char' } },
+        finish_reason: null
+      }
+    ]
+  },
+  {
+    id: 'chatcmpl-7WBy19k4tnzMa0svAIAqkqeIaKZh8',
+    object: 'chat.completion.chunk',
+    created: 1687906853,
+    model: 'gpt-3.5-turbo-0613',
+    choices: [
+      {
+        index: 0,
+        delta: { function_call: { arguments: 'l' } },
+        finish_reason: null
+      }
+    ]
+  },
+  {
+    id: 'chatcmpl-7WBy19k4tnzMa0svAIAqkqeIaKZh8',
+    object: 'chat.completion.chunk',
+    created: 1687906853,
+    model: 'gpt-3.5-turbo-0613',
+    choices: [
+      {
+        index: 0,
+        delta: { function_call: { arguments: 'ottesville' } },
+        finish_reason: null
+      }
+    ]
+  },
+  {
+    id: 'chatcmpl-7WBy19k4tnzMa0svAIAqkqeIaKZh8',
+    object: 'chat.completion.chunk',
+    created: 1687906853,
+    model: 'gpt-3.5-turbo-0613',
+    choices: [
+      {
+        index: 0,
+        delta: { function_call: { arguments: ',' } },
+        finish_reason: null
+      }
+    ]
+  },
+  {
+    id: 'chatcmpl-7WBy19k4tnzMa0svAIAqkqeIaKZh8',
+    object: 'chat.completion.chunk',
+    created: 1687906853,
+    model: 'gpt-3.5-turbo-0613',
+    choices: [
+      {
+        index: 0,
+        delta: { function_call: { arguments: ' Virginia' } },
+        finish_reason: null
+      }
+    ]
+  },
+  {
+    id: 'chatcmpl-7WBy19k4tnzMa0svAIAqkqeIaKZh8',
+    object: 'chat.completion.chunk',
+    created: 1687906853,
+    model: 'gpt-3.5-turbo-0613',
+    choices: [
+      {
+        index: 0,
+        delta: { function_call: { arguments: '",\n' } },
+        finish_reason: null
+      }
+    ]
+  },
+  {
+    id: 'chatcmpl-7WBy19k4tnzMa0svAIAqkqeIaKZh8',
+    object: 'chat.completion.chunk',
+    created: 1687906853,
+    model: 'gpt-3.5-turbo-0613',
+    choices: [
+      {
+        index: 0,
+        delta: { function_call: { arguments: '"' } },
+        finish_reason: null
+      }
+    ]
+  },
+  {
+    id: 'chatcmpl-7WBy19k4tnzMa0svAIAqkqeIaKZh8',
+    object: 'chat.completion.chunk',
+    created: 1687906853,
+    model: 'gpt-3.5-turbo-0613',
+    choices: [
+      {
+        index: 0,
+        delta: { function_call: { arguments: 'format' } },
+        finish_reason: null
+      }
+    ]
+  },
+  {
+    id: 'chatcmpl-7WBy19k4tnzMa0svAIAqkqeIaKZh8',
+    object: 'chat.completion.chunk',
+    created: 1687906853,
+    model: 'gpt-3.5-turbo-0613',
+    choices: [
+      {
+        index: 0,
+        delta: { function_call: { arguments: '":' } },
+        finish_reason: null
+      }
+    ]
+  },
+  {
+    id: 'chatcmpl-7WBy19k4tnzMa0svAIAqkqeIaKZh8',
+    object: 'chat.completion.chunk',
+    created: 1687906853,
+    model: 'gpt-3.5-turbo-0613',
+    choices: [
+      {
+        index: 0,
+        delta: { function_call: { arguments: ' "' } },
+        finish_reason: null
+      }
+    ]
+  },
+  {
+    id: 'chatcmpl-7WBy19k4tnzMa0svAIAqkqeIaKZh8',
+    object: 'chat.completion.chunk',
+    created: 1687906853,
+    model: 'gpt-3.5-turbo-0613',
+    choices: [
+      {
+        index: 0,
+        delta: { function_call: { arguments: 'c' } },
+        finish_reason: null
+      }
+    ]
+  },
+  {
+    id: 'chatcmpl-7WBy19k4tnzMa0svAIAqkqeIaKZh8',
+    object: 'chat.completion.chunk',
+    created: 1687906853,
+    model: 'gpt-3.5-turbo-0613',
+    choices: [
+      {
+        index: 0,
+        delta: { function_call: { arguments: 'elsius' } },
+        finish_reason: null
+      }
+    ]
+  },
+  {
+    id: 'chatcmpl-7WBy19k4tnzMa0svAIAqkqeIaKZh8',
+    object: 'chat.completion.chunk',
+    created: 1687906853,
+    model: 'gpt-3.5-turbo-0613',
+    choices: [
+      {
+        index: 0,
+        delta: { function_call: { arguments: '"\n' } },
+        finish_reason: null
+      }
+    ]
+  },
+  {
+    id: 'chatcmpl-7WBy19k4tnzMa0svAIAqkqeIaKZh8',
+    object: 'chat.completion.chunk',
+    created: 1687906853,
+    model: 'gpt-3.5-turbo-0613',
+    choices: [
+      {
+        index: 0,
+        delta: { function_call: { arguments: '}' } },
+        finish_reason: null
+      }
+    ]
+  },
+  {
+    id: 'chatcmpl-7WBy19k4tnzMa0svAIAqkqeIaKZh8',
+    object: 'chat.completion.chunk',
+    created: 1687906853,
+    model: 'gpt-3.5-turbo-0613',
+    choices: [{ index: 0, delta: {}, finish_reason: 'function_call' }]
   }
 ]