+++ packages/core/tests/utils/mock-service.ts
@@ -1,6 +1,9 @@
 import { ServerResponse, createServer } from 'node:http'
 
-import Snapshot_OpenAIChat from '../snapshots/openai-chat'
+import {
+  chatCompletionChunks,
+  chatCompletionChunksWithFunctionCall
+} from '../snapshots/openai-chat'
 
 async function flushDataToResponse(
   res: ServerResponse,
@@ -37,9 +40,40 @@ export const setup = () => {
 
   const server = createServer((req, res) => {
     const service = req.headers['x-mock-service'] || 'openai'
-    const type = req.headers['x-mock-type'] || 'chat'
+    const type = req.headers['x-mock-type'] || 'chat' || 'func_call'
 
     switch (type) {
+      case 'func_call': // new case
+        switch (service) {
+          case 'openai':
+            res.writeHead(200, {
+              'Content-Type': 'text/event-stream',
+              'Cache-Control': 'no-cache',
+              Connection: 'keep-alive'
+            })
+            res.flushHeaders()
+            recentFlushed = []
+            flushDataToResponse(
+              res,
+              chatCompletionChunksWithFunctionCall.map(
+                value =>
+                  new Proxy(
+                    { value },
+                    {
+                      get(target) {
+                        recentFlushed.push(target.value)
+                        return target.value
+                      }
+                    }
+                  )
+              ),
+              '[DONE]'
+            )
+            break
+          default:
+            throw new Error(`Unknown service: ${service}`)
+        }
+        break
       case 'chat':
         switch (service) {
           case 'openai':
@@ -52,7 +86,7 @@ export const setup = () => {
             recentFlushed = []
             flushDataToResponse(
               res,
-              Snapshot_OpenAIChat.map(
+              chatCompletionChunks.map(
                 value =>
                   new Proxy(
                     { value },